+++ src/app/components/compare_widget/compare_widget.directive.js
@@ -13,7 +13,6 @@
             },
             controller: 'CompareWidgetController',
             controllerAs: 'vm',
-            restrict: 'EA',
             scope: {},
         };
     }